+++ packages/widgetbook/lib/src/addons/device_frame_addon/device_frame_addon.dart
@@ -81,11 +81,20 @@ class DeviceFrameAddon extends WidgetbookAddon<DeviceFrameSetting> {
           orientation: setting.orientation,
           device: setting.device,
           isFrameVisible: setting.hasFrame,
-          screen: setting.hasFrame
-              ? child
-              : SafeArea(
-                  child: child,
-                ),
+          // A navigator below the device frame is necessary to make
+          // the popup routes (e.g. dialogs and bottom sheets) work within
+          // the device frame, otherwise they would use the navigator from
+          // the app builder, causing these routes to fill the whole
+          // workbench and not just the device frame.
+          screen: Navigator(
+            onGenerateRoute: (_) => PageRouteBuilder(
+              pageBuilder: (context, _, __) => setting.hasFrame
+                  ? child
+                  : SafeArea(
+                      child: child,
+                    ),
+            ),
+          ),
         ),
       ),
     );